Enable relative dates based off waffle
With this change, the api will only return relative block dates
and/or relative user override dates if a waffle flag is enabled
for the given course and user.

Also, remove the unused and inaccurate api decorator
override_enabled. It referenced an old waffle flag that gated all
of edx-when. Due to this change, I've bumped the version to 1.0.0.

AA-27