Cache ContentDate data in external cache (v1.1.3)
The get_dates_for_course function makes large queries to fetch
ContentDate objects. Since this function is called from almost
any courseware operation (we fetch a whole course worth of dates
at a time when doing XBlock operations), it started putting a
lot of stress on the database under load. In order to help
alleviate some of that pressure, this commit uses memcached to
cache this data.

Ideally, we would be smarter about only grabbing the data we
actually need, rather than grabbing the entire course's date
information at once. Also, the fact that we're just shoving
model objects into the cache means that this functionality relies
on using pickle caching serialization rather than JSON, which is
worse from a security and compatibility point of view. This is not
great code, it's just the simplest/least-risky thing we can do to
get to scale the site in a hurry.