Drop ContentDate.active database index (v1.1.4)
Queries for ContentDate for a given course on Aurora MySQL were
using an index_merge on the course_id and active fields. We believe
this was causing a lot of extra work because almost all rows are
active, and it was therefore doing an intersect on the useful
course_id index and a nearly useless but increasingly growing active
index.

This commit drops the active index entirely so that it will use only
the course_id index. We also bump our version to 1.1.4.