AA-195: Don't give rel dates past cutoff date
The cutoff date is the point past which the learner will not have
time to complete the course, because expected due dates would be
due past the course end date.

This new check only considers if the learner's original schedule
was past the cutoff, it does not consider schedules that have been
reset. (i.e. we still allow for learners that reset their schedule
close to the course end date, by clustering their due dates at
course end).